reuse repr methods as mixin
