+++ pyroll/core/__init__.py
@@ -1,3 +1,5 @@
+from . import shapes as _
+
 from .solve import solve
 from .grooves import BoxGroove, ConstrictedBoxGroove, SquareGroove, DiamondGroove, RoundGroove, FalseRoundGroove, \
     CircularOvalGroove, FlatOvalGroove, SwedishOvalGroove, ConstrictedSwedishOvalGroove, Oval3RadiiGroove, \
